repackage : Application 계층 내 usecase, command 분리
