--- src/main/java/tdd/groomingzone/board/freeboard/application/port/in/DeleteFreeBoardUseCase.java
@@ -1,6 +1,0 @@
-package tdd.groomingzone.board.freeboard.application.port.in;
-
-public interface DeleteFreeBoardUseCase {
-
-    void deleteFreeBoard(DeleteFreeBoardCommand command);
-}
