--- src/main/java/tdd/groomingzone/board/freeboard/application/port/in/PostFreeBoardUseCase.java
@@ -1,6 +1,0 @@
-package tdd.groomingzone.board.freeboard.application.port.in;
-
-public interface PostFreeBoardUseCase {
-
-    FreeBoardCommandResponse postFreeBoard(PostFreeBoardCommand command);
-}
