--- src/main/java/tdd/groomingzone/board/freeboard/application/port/in/PutFreeBoardUseCase.java
@@ -1,6 +1,0 @@
-package tdd.groomingzone.board.freeboard.application.port.in;
-
-public interface PutFreeBoardUseCase {
-
-    FreeBoardCommandResponse putFreeBoard(PutFreeBoardCommand putFreeBoardCommand);
-}
