+++ src/main/java/tdd/groomingzone/board/freeboard/application/port/in/usecase/DeleteFreeBoardUseCase.java
@@ -1,0 +1,8 @@
+package tdd.groomingzone.board.freeboard.application.port.in.usecase;
+
+import tdd.groomingzone.board.freeboard.application.port.in.command.DeleteFreeBoardCommand;
+
+public interface DeleteFreeBoardUseCase {
+
+    void deleteFreeBoard(DeleteFreeBoardCommand command);
+}
