+++ src/main/java/tdd/groomingzone/board/freeboard/application/port/in/usecase/PutFreeBoardUseCase.java
@@ -1,0 +1,9 @@
+package tdd.groomingzone.board.freeboard.application.port.in.usecase;
+
+import tdd.groomingzone.board.freeboard.application.port.in.FreeBoardCommandResponse;
+import tdd.groomingzone.board.freeboard.application.port.in.command.PutFreeBoardCommand;
+
+public interface PutFreeBoardUseCase {
+
+    FreeBoardCommandResponse putFreeBoard(PutFreeBoardCommand putFreeBoardCommand);
+}
