--- src/main/java/tdd/groomingzone/member/application/port/in/PostMemberUseCase.java
@@ -1,6 +1,0 @@
-package tdd.groomingzone.member.application.port.in;
-
-public interface PostMemberUseCase {
-
-    MemberCommandResponse postMember(PostMemberCommand command);
-}
