+++ src/main/java/tdd/groomingzone/member/application/service/PostMemberService.java
@@ -6,8 +6,8 @@ import org.springframework.transaction.annotation.Transactional;
 import tdd.groomingzone.global.exception.BusinessException;
 import tdd.groomingzone.global.exception.ExceptionCode;
 import tdd.groomingzone.member.application.port.out.LoadMemberPort;
-import tdd.groomingzone.member.application.port.in.PostMemberUseCase;
-import tdd.groomingzone.member.application.port.in.PostMemberCommand;
+import tdd.groomingzone.member.application.port.in.usecase.PostMemberUseCase;
+import tdd.groomingzone.member.application.port.in.command.PostMemberCommand;
 import tdd.groomingzone.member.application.port.in.MemberCommandResponse;
 import tdd.groomingzone.member.application.port.out.SaveMemberPort;
 import tdd.groomingzone.member.domain.Member;
